feat(api-text-porcessing): define API predict endpoint
